updated gateway projects to split user and admin routes
