+++ items/gateway_svc/apis/web/admin/__init__.py
@@ -1,0 +1,30 @@
+"""
+Copyright 2025 Integrated Test Management Suite Development Team
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+import logging
+import quart
+from sessions import Sessions
+from .projects_api import create_blueprint as create_projects_bp
+
+
+def create_web_admin_routes(logger: logging.Logger,
+                            sessions: Sessions):
+    web_bp = quart.Blueprint("web_admin_routes", __name__)
+
+    # Projects API routes
+    web_bp.register_blueprint(create_projects_bp(logger))
+
+    return web_bp
+
