+++ items/gateway_svc/apis/web/projects_api_view.py
@@ -87,91 +87,3 @@ class ProjectsApiView(BaseView):
         return quart.Response(json.dumps(api_response.body),
                               status=http.HTTPStatus.OK,
                               content_type="application/json")
-
-    @validate_json(json_schemas.SCHEMA_ADD_PROJECT_REQUEST)
-    async def add_project(self, request_msg: ApiResponse):
-
-        cms_svc: str = ThreadSafeConfiguration().apis_cms_svc
-        url: str = f"{cms_svc}projects/add"
-
-        cms_request: dict = {
-            "name": request_msg.body.name,
-            "announcement": request_msg.body.announcement,
-            "announcement_on_overview":
-                request_msg.body.announcement_on_overview,
-        }
-        api_response = await self._call_api_post(url, cms_request)
-
-        if api_response.status_code != http.HTTPStatus.OK:
-            self._logger.critical("CMS SVC /projects/add request invalid"
-                                  " - Reason: %s",api_response.exception_msg)
-            response_json = {
-                "status": 0,
-                'error': api_response.body['error_msg']
-            }
-            return quart.Response(json.dumps(response_json),
-                                  status=http.HTTPStatus.BAD_REQUEST,
-                                  content_type="application/json")
-
-        response_json: dict = {"status": 1}
-        return quart.Response(json.dumps(response_json),
-                              status=http.HTTPStatus.OK,
-                              content_type="application/json")
-
-    @validate_json(json_schemas.SCHEMA_MODIFY_PROJECT_REQUEST)
-    async def modify_project(self,
-                             request_msg: ApiResponse,
-                             project_id: int):
-        cms_svc: str = ThreadSafeConfiguration().apis_cms_svc
-        url: str = f"{cms_svc}projects/modify/{project_id}"
-
-        cms_request: dict = {
-            "name": request_msg.body.name,
-            "announcement": request_msg.body.announcement,
-            "announcement_on_overview":
-                request_msg.body.announcement_on_overview,
-        }
-        api_response = await self._call_api_post(url, cms_request)
-
-        if api_response.status_code != http.HTTPStatus.OK:
-            self._logger.critical("CMS SVC /projects/modify request invalid"
-                                  " - Reason: %s",api_response.exception_msg)
-            response_json = {
-                "status": 0,
-                'error': api_response.body['error_msg']
-            }
-            return quart.Response(json.dumps(response_json),
-                                  status=http.HTTPStatus.BAD_REQUEST,
-                                  content_type="application/json")
-
-        response_json: dict = {"status": 1}
-        return quart.Response(json.dumps(response_json),
-                              status=http.HTTPStatus.OK,
-                              content_type="application/json")
-
-    async def delete_project(self, project_id: int):
-        cms_svc: str = ThreadSafeConfiguration().apis_cms_svc
-        url: str = f"{cms_svc}projects/delete/{project_id}?hard_delete=true"
-
-        api_response = await self._call_api_delete(url)
-
-        if api_response.status_code == http.HTTPStatus.BAD_REQUEST:
-            return quart.Response(json.dumps(api_response.body),
-                                  status=http.HTTPStatus.BAD_REQUEST,
-                                  content_type="application/json")
-
-        if api_response.status_code != http.HTTPStatus.OK:
-            self._logger.critical(
-                "CMS SVC %s request invalid - Reason: %s",
-                url, api_response.exception_msg)
-            response_json = {
-                "status": 0,
-                'error': 'Internal error!'
-            }
-            return quart.Response(json.dumps(response_json),
-                                  status=http.HTTPStatus.INTERNAL_SERVER_ERROR,
-                                  content_type="application/json")
-
-        return quart.Response(json.dumps(api_response.body),
-                              status=http.HTTPStatus.OK,
-                              content_type="application/json")
